requirements.txt
streamlit
pillow
opencv-python
scikit-learn
numpy
matplotlib
